[init] Adds the check for superfluous file existence.

+++ bin/initial_package_install.sh
@@ -13,7 +13,9 @@ sudo apt-get install -y apt-transport-https ca-certificates \
 wget -q -O - https://dl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
 sudo sh -c 'echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" > /etc/apt/sources.list.d/google.list'
 sudo sh -c 'echo "deb [arch=amd64] http://dl.google.com/linux/talkplugin/deb/ stable main" >> /etc/apt/sources.list.d/google.list'
-sudo rm /etc/apt/sources.list.d/google-chrome-beta.list
+if [ -e /etc/apt/sources.list.d/google-chrome-beta.list ]; then
+  sudo rm /etc/apt/sources.list.d/google-chrome-beta.list
+fi
 
 # Docker repo.
 curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
